add formatting, conform to some best practices
For the example, I expand struct arguments so it is more visible on what is needed and so it does not hard wrap, causing it to become unreadable
If you insist on making a package with tonnes of structs and structs in those structs, please do this in your examples, else it is nearly impossible to read and/or learn what your code is doing.

Can you provide some input on the ordering (and significance of said ordering) of AddYAxis() and AddDataset()? Does it matter?